requirements.txt
opencv-python==4.5.5.64
numpy==1.22.3
PyAutoGUI==0.9.53
pytesseract==0.3.9
Pillow==9.1.0